Implement gradient operator

+++ C++/data_types/grid.cpp
@@ -12,9 +12,9 @@ namespace cfd
 namespace geometry
 {
 
-std::uint64_t Grid::GetSize() const
+std::int32_t Grid::GetSize() const
 {
-    return elements_.size();
+    return static_cast<std::int32_t>(elements_.size());
 }
 
 const std::vector<std::optional<double>> Node::GetValues() const
